+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Features
 
+- Export function for network and logging
 - Add logrus formatter
 - Integrate urfave instead of envconfig
 - Add log level for plugins